pyproject.toml
[project]
name = "overengineered-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-environ>=0.12.0",
    "django>=5.2.4",
    "pre-commit>=4.2.0",
    "django-storages>=1.14.6",
    "pillow>=11.3.0",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "ruff>=0.12.5",
]
prod = [
    "gunicorn>=23.0.0",
]
